moved functions to AESmath
